moving es and db accessors
