+++ corehq/apps/reports/analytics/esaccessors.py
@@ -212,6 +212,16 @@ def get_last_form_submissions_by_user(domain, user_ids, app_id=None, xmlns=None)
     return result
 
 
+def get_last_form_for_apps(apps, user_id):
+    forms = []
+    for app_id in apps:
+        query = FormES().app(app_id).user_id(user_id).sort('received_on', desc=True).size(1)
+        hits = query.run().hits
+        if hits:
+            forms.append(hits[0])
+    return forms
+
+g
 def get_submission_counts_by_user(domain, datespan, user_ids=None):
     return _get_form_counts_by_user(domain, datespan, True, user_ids)
